refactor: move iam-specific functionality to a base class
This will allow other authenticators that interact with the IAM
service to share configuration and functionality.
